+++ src/index.tsx
@@ -4,7 +4,7 @@ import './index.css';
 
 import { Provider } from 'react-redux';
 
-import configureStore from 'state/configure-store';
+import configureStore from 'store/configure-store';
 import AppImpl from 'renderer/components/app-impl';
 
 import * as serviceWorker from './serviceWorker';
