+++ src/library/activity-parsers/shared-structures.ts
