+++ src/library/activity-parsers/xml-helpers.ts
